migrate API functionality into experiment-context
