+++ .env.development
@@ -1,0 +1,1 @@
+NEXT_API_SERVER=http://localhost:9090/v1.0
