+++ context/experiment-context.tsx
@@ -1,9 +1,15 @@
 import * as React from 'react'
 import { versionInfo } from '../components/version-info'
 import { useLocalStorageReducer } from '../hooks/useLocalStorageReducer'
+import {
+  Configuration,
+  DefaultApi,
+  OptimizerapiOptimizerRunRequest,
+} from '../openapi'
 import { Dispatch, rootReducer } from '../reducers/reducers'
 import { initialState, State } from '../store'
 import { ExperimentResultType, ExperimentType } from '../types/common'
+import { calculateData, calculateSpace } from '../utility/converters'
 import { migrate } from '../utility/migration/migration'
 import { useGlobal } from './global-context'
 
@@ -73,12 +79,55 @@ function useExperiment() {
   return context
 }
 
+const fetchExperimentResult = async (
+  experiment: ExperimentType
+): Promise<ExperimentResultType> => {
+  const API_SERVER = process.env.NEXT_API_SERVER
+  const api = new DefaultApi(new Configuration({ basePath: API_SERVER }))
+  const cfg = experiment.optimizerConfig
+  const extras = experiment.extras || {}
+  const space = calculateSpace(experiment)
+
+  const request: OptimizerapiOptimizerRunRequest = {
+    experiment: {
+      data: calculateData(
+        experiment.categoricalVariables,
+        experiment.valueVariables,
+        experiment.scoreVariables,
+        experiment.dataPoints
+      ),
+      extras: extras,
+      optimizerConfig: {
+        acqFunc: cfg.acqFunc,
+        baseEstimator: cfg.baseEstimator,
+        initialPoints: Number(cfg.initialPoints),
+        kappa: Number(cfg.kappa),
+        xi: Number(cfg.xi),
+        space: space,
+      },
+    },
+  }
+
+  const result = await api.optimizerapiOptimizerRun(request)
+
+  const experimentResult: ExperimentResultType = {
+    id: experiment.id,
+    plots:
+      result.plots &&
+      result.plots.map(p => {
+        return { id: p.id, plot: p.plot }
+      }),
+    next: result.result.next,
+    pickled: result.result.pickled,
+    expectedMinimum: result.result.models.find(() => true)?.expectedMinimum,
+    extras: result.result.extras,
+  }
+
+  return experimentResult
+}
+
 async function runExperiment(dispatch: Dispatch, experiment: ExperimentType) {
-  const response: Response = await fetch(`/api/experiment/${experiment.id}`, {
-    method: 'POST',
-    body: JSON.stringify(experiment),
-  })
-  const result: ExperimentResultType = await response.json()
+  const result = await fetchExperimentResult(experiment)
   dispatch({ type: 'registerResult', payload: result })
 }
