--- pages/api/experiment/[id].test.tsx
@@ -1,7 +1,0 @@
-export {}
-
-describe('/api/experiment/[id]', () => {
-  test('POST calls API backend', () => {
-    // TODO Test is not implemented
-  })
-})
